Text handler tests passed
problems: text handler does not flow propperly. if anyone wants to fix
it talk to me and i'll tell you how

still to do: future release of the text handler to handle flowing text
properly

+++ SmartTrolley/src/graphicshandler/SlideElementDuration.java
@@ -39,8 +39,8 @@ public class SlideElementDuration {
 
 	private Timer timer;
 	private Node node;
-	private int duration = 0;
-	private int startTime = 0;
+	private double duration = 0;
+	private double startTime = 0;
 	private boolean started;
 
 	
@@ -75,7 +75,7 @@ public class SlideElementDuration {
 			started = true;
 
 			TimerTask appear = new ShowTask();
-			timer.schedule(appear, this.startTime);
+			timer.schedule(appear, (long)this.startTime);
 		}
 	}
 
@@ -86,7 +86,7 @@ public class SlideElementDuration {
 	*<p> Date Modified: 25 April 2014
 	
 	*/
-	public void setStartTime(int seconds) {
+	public void setStartTime(double seconds) {
 		
 		//startTime is in milliseconds
 		startTime = seconds*1000;
@@ -100,7 +100,7 @@ public class SlideElementDuration {
 	*<p> Date Modified: 25 April 2014
 	
 	*/
-	public void setDuration(int seconds) {
+	public void setDuration(double seconds) {
 		// duration is in milliseconds
 		duration = seconds * 1000;
 	}
@@ -121,7 +121,7 @@ public class SlideElementDuration {
 
 				// 0 duration keeps shape visible
 				if (duration != 0) {
-					timer.schedule(new ShowTask(), duration);
+					timer.schedule(new ShowTask(), (long)duration);
 				}
 
 			} else {
